+++ doc/source/changelog.rst
@@ -4,6 +4,20 @@ Changelog
 All notable changes to this project will be documented in this file.  This
 project adheres to `Semantic Versioning <http://semver.org/spec/v2.0.0.html>`_.
 
+Version 2.0.0
+-------------
+
+New:
+
+  * Routines to compute contribution matrices have been added.
+  * Further documentation on the rates retrieval method has been added.
+
+Fix:
+
+  * The GrRproc class now automatically detects the maximum number of delayed
+    neutrons, so the *n_bdn_max* optional keyword has been removed.  This is
+    a backwards-incompatible change.
+
 Version 1.0.1
 -------------